Loop in gRPC read call until buffer is full
Before this change the buffer was often getting only a little bit of
data before returning, leading to dramatically more read calls than
necessary, and hurting throughput.

Looping until the buffer fills up nearly doubled throughput in the test
cases for the TestDFSIO benchmark, and cut the number of `read` calls
by a factor of 10 or more.

Change-Id: I2b927acd47c7e851c2664d3ffba1edc440fd27ba